--- pets-lib/src/dialogue/dbox/choice.rs
@@ -1,63 +1,0 @@
-//!
-//! Stuff specifically related to dialog box choice labels
-//!
-
-use dialogical::DialogueChoice;
-
-use godot::engine::RichTextLabel;
-use godot::prelude::*;
-
-use crate::consts::dialogue::*;
-use crate::prelude::*;
-
-use super::dchoice::DChoice;
-use super::DialogBox;
-
-impl DialogBox {
-    /// delete old labels and create new default ones
-    pub(super) fn recreate_choice_labels(&mut self, choices: &[DialogueChoice]) {
-        self.free_choice_labels();
-
-        let mut container = self.choice_container();
-
-        for (i, choice) in choices.iter().enumerate() {
-            let dchoice = DChoice::new_container(i, &choice.text);
-            container.add_child(dchoice.upcast());
-        }
-    }
-
-    pub(super) fn tween_choices_wave(&mut self, up: bool) {
-        for (i, cont) in self.choice_labels().iter_shared().enumerate() {
-            // if moving up, start below the window
-            if up {
-                godot_print!("Up!");
-                cont.get_node_as::<RichTextLabel>("Label")
-                    .set_position(Vector2::new(0.0, DBOX_CHOICE_HEIGHT));
-            } else {
-                godot_print!("Down!");
-            }
-
-            // we can't move the label into the closure because of
-            // thread safety stuff, so just pass in the instance id
-            let label_id = cont.instance_id();
-
-            let func = Callable::from_fn("choice_slide_up", move |_| {
-                // get the label again using the instance id
-                let label = Gd::<DChoice>::try_from_instance_id(label_id);
-
-                let Ok(label) = label else {
-                    return Ok(Variant::from(()));
-                };
-
-                label.bind().tween_label(up);
-                Ok(Variant::from(()))
-            });
-
-            // set timer
-            godot_tree()
-                .create_timer(DBOX_CHOICE_WAVE_TIME * i as f64)
-                .unwrap()
-                .connect("timeout".into(), func);
-        }
-    }
-}
